requirements.txt
numpy
tsplib95
grpcio
grpcio_tools
deap
